# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Food Delivery API",
        "description": "Food Delivery API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/drivers": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Get list of drivers",
                "operationId": "90f1e2d7ba23baa1e136288ed235666c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Driver"
                                    }
                                }
                            }
                        }
                    }
                },
                "servers": [
                    {
                        "url": "/"
                    }
                ]
            },
            "post": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Create a new driver",
                "operationId": "6bdf167696d642223e0d27e88a0b4896",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreDriverRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Driver"
                                }
                            }
                        }
                    }
                },
                "servers": [
                    {
                        "url": "/"
                    }
                ]
            }
        },
        "/api/drivers/all": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Get all drivers with cursor pagination (for bulk fetch)",
                "operationId": "176d6eff22ca42ff68a6ff9125c93aad",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 500
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Driver"
                                    }
                                }
                            }
                        }
                    }
                },
                "servers": [
                    {
                        "url": "/"
                    }
                ]
            }
        },
        "/api/drivers/{id}": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Get driver details",
                "operationId": "1de6dd3c4dadedd666e6f9e425fd3db4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Driver"
                                }
                            }
                        }
                    }
                },
                "servers": [
                    {
                        "url": "/"
                    }
                ]
            },
            "put": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Update an existing driver",
                "operationId": "de2a8f73505116569bbec3fa525db6bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDriverRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Driver"
                                }
                            }
                        }
                    }
                },
                "servers": [
                    {
                        "url": "/"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Delete a driver",
                "operationId": "c32adc70b84b5c4d5f214e60757aaee3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                },
                "servers": [
                    {
                        "url": "/"
                    }
                ]
            }
        },
        "/api/drivers/available": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Get available drivers",
                "operationId": "cd03eca28abe5b326e6e0ea68287161c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Driver"
                                    }
                                }
                            }
                        }
                    }
                },
                "servers": [
                    {
                        "url": "/"
                    }
                ]
            }
        },
        "/api/drivers/{id}/history": {
            "get": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Get driver delivery history",
                "operationId": "2a86ad982a65dfc5ef877727ba2f80c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DeliveryHistory"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "servers": [
                    {
                        "url": "/"
                    }
                ]
            }
        },
        "/api/drivers/login": {
            "post": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Driver Login",
                "operationId": "4d92224ab9bed2c95c2be67479380a93",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DriverLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "servers": [
                    {
                        "url": "/"
                    }
                ]
            }
        },
        "/api/drivers/register": {
            "post": {
                "tags": [
                    "Drivers"
                ],
                "summary": "Driver Registration",
                "operationId": "4f7e14ac48a56c872583a6bf09fcc9f9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DriverRegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "driver": {
                                            "$ref": "#/components/schemas/Driver"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "servers": [
                    {
                        "url": "/"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Driver": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Budi Santoso"
                    },
                    "email": {
                        "type": "string",
                        "example": "budi@driver.com"
                    },
                    "vehicle": {
                        "type": "string",
                        "example": "Honda Vario B 1234 XYZ"
                    },
                    "status": {
                        "type": "string",
                        "example": "available"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-21T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-21T10:00:00Z"
                    }
                },
                "type": "object"
            },
            "StoreDriverRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "vehicle",
                    "status"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Andi Saputra"
                    },
                    "email": {
                        "type": "string",
                        "example": "andi@driver.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    },
                    "vehicle": {
                        "type": "string",
                        "example": "Yamaha Beat B 5678 GHI"
                    },
                    "status": {
                        "type": "string",
                        "example": "available"
                    }
                },
                "type": "object"
            },
            "UpdateDriverRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Andi Saputra"
                    },
                    "vehicle": {
                        "type": "string",
                        "example": "Yamaha Beat B 5678 GHI"
                    },
                    "status": {
                        "type": "string",
                        "example": "unavailable"
                    }
                },
                "type": "object"
            },
            "DriverLoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "budi@driver.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
            "DriverRegisterRequest": {
                "required": [
                    "name",
                    "email",
                    "password",
                    "vehicle"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Cecep Kuncoro"
                    },
                    "email": {
                        "type": "string",
                        "example": "cecep@driver.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "password123"
                    },
                    "vehicle": {
                        "type": "string",
                        "example": "Suzuki NMax D 7642 TTR"
                    }
                },
                "type": "object"
            },
            "DeliveryHistory": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "driver_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "order_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "delivered_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-21T11:00:00Z"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-21T11:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-21T11:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}